[REFACTOR] Created class to encapsulate common test needs [LUNN/BARREYRE]
